bkr-create-dummy-datajs : created.
