+++ backend/.gitignore
@@ -2,3 +2,4 @@ node_modules
 .env
 test.rest
 requests.http
+data
