+++ backend/index.js
@@ -6,7 +6,7 @@ import path from "path";
 //database on mongodb updated : 4 Ocober 2022, 11:25
 
 // import session from "express-session";
-import cookieParser from "cookie-parser";
+// import cookieParser from "cookie-parser";
 
 import userRouter from "./routes/userRoutes.js";
