+++ backend/routes/seedRoutes.js
@@ -1,6 +1,6 @@
 import express from "express";
 import Product from "../models/productModel.js";
-import data from "../data.js";
+import data from "../data/data.js";
 import User from "../models/userModel.js";
 
 const seedRouter = express.Router();
